python/requirements.txt
numpy
sklearn
tensorflow
seaborn
matplotlib
regex
nltk
flask
gensim
pyyaml